journey_core_nlp/requirements.txt
kafka-python
requests